search: expose GitHub/GitLab topics in UI (#58927)
Currently, we sync topics from GitHub and GitLab and it is possible to
filter repositories by their topics with the `repo:has.topic` filter.

However, unlike user-added metadata, the synced topics are not visible in the UI,
neither for repo matches nor on the repo tree page. This means it is impossible 
for users to figure out which topics they can filter by.

With this PR we display topics, such as "language" or "golang", in addition to 
user-added metadata for repo matches and on the repo tree page.

## Test plan

Note: Our search language distinguishes between user provided metadata (has.meta) and automatically synced metadata (has.topic). 

- manual testing
  - I checked that clicking on a "topic" badge adds a "has.topic" filter, while clicking on a "metadata" badge add a "has.meta" filter
  - Adding and deleting metadata works like before
  - Topics and metadata are sorted alphabetically
  - Results without metadata or topics are displayed correctly
- added new unit test