search: support "exact match" pattern using "..."
Relates to #58815

With this change, a quoted pattern, like in `"foo bar"`, is interpreted
literally IE spaces are interpreted as spaces instead of as logical AND.
Quotes, which are part of the pattern have to be escaped.

Example:

searching for the literal `foo "bar"`, where "bar" is surrounded by quotes, the
query is `"foo \"bar\""` (or equivalently `'foo "bar"'`)

Note: This only applies for our keyword search prototype.

Test plan:
- updated unit test
- manual testing
  - tried out various combinations