Searcher: pull out 'matcher' interface (#59040)
Currently searcher can only handle patterns with a single atom. We'd like to update searcher to handle AND/ OR patterns directly.

This refactor takes us a step towards being able to natively handle AND/ OR patterns by refactoring the way we match files. It pulls out a `matcher` interface, which for now just has a single implementation (`regexMatcher`), but eventually can represent nested matchers combined through AND/ OR. Even without this goal, I think the refactor stands on its own and helps clean up the logic.

Specific changes:
* Lots of renames to make things clearer, like `readerGrep` -> `regexMatcher`
* Define `matcher` interface, which finds matches given file contents, and related logic into new file
* Pull `pathMatcher` out of `matcher`, since it's really its own top-level concept
* Move all file content and buffer management to `search_regex`. This lets each `matcher` object become threadsafe, and avoids the need for copying them for every goroutine.