svelte: Fix setting theme from temporary settings (#59091)
Turns out the theme values stored in temporary setting are different
from what the prototype expected. It only accidentally worked when I
tested it.

This commit introduces a helper method that accepts an arbitrary string
value and updates the theme state accordingly, instead of using an
arbitrary string value as theme state (ideally TS would have complained
about this before but it looks like it didn't).