svelte: Fix local build (#59100)
I don't know if this is only an issue for me, but I'm not able to run pnpm build locally anymore. The build fails with the error mentioned in the comment. Hover seeing that CI is green for other PRs and SvelteKit is deployed to S2 successfully, I don't really know why this is happening.
This change fixes it though.