sg: include MSP toolchain in default builds (#59087)
This change bundles the `sg msp` toolchain in default `sg` builds. This was previously in a very experimental state and introduced a significant increase in the `sg` binary size, so we had it behind a build flag. However, we've discussed this a few times with Dev Infra and the consensus is that it is okay to include by default.

We have a demo prepared for this week, and with many of our core features now available, we want to make a renewed push for adoption, so now is a good time to start bundling this command by default.

There are no functionality changes - this PR just removes the overwrite-command-on-init-with-build-flag stuff and exports the full toolchain by default from the `dev/sg/msp` package.

Close https://github.com/sourcegraph/managed-services/issues/182

## Test plan

```
➜ go build -o ./sg ./dev/sg                                 
➜ ./sg msp init -h
NAME:
   sg managed-services-platform init - Initialize a template Managed Services Platform service spec

USAGE:
   
   sg msp init -owner core-services -name "MSP Example Service" msp-example


OPTIONS:
   --kind value   Kind of service (one of: 'service', 'job') (default: "service")
   --owner value  Name of team owning this new service
   --name value   Specify a human-readable name for this service
   --dev          Generate a dev environment as the initial environment (d
```