Svelte: highlight expanded lines (#59101)
When clicking the "expand results" button, the expanded lines were not syntax highlighted for the svelte prototype. This happened the ranges we were fetching for highlights are pre-calculated from the collapsed groups. Instead, we want to fetch ranges for every group so when we expand, we have the highlighted result already. The first commit does that.