Remove App from codebase
Cody no longer needs it and it is obsolete now!

Since App added a non-insignificant amount of new concepts and alternative code paths, I decided to take some time and remove it from out codebase.

So.. here we go.

This should not affect any of the existing deployments, we only remove functionality that was special-cased for app.