msp: fixed-length operator access account ID, restrict service ID lengths (#59187)
Using service name in the account ID results in IDs that are too long for the operator access SA added in #59105 - this change uses a randomized name with fixed length. This fits the use case as well, since these operator accounts should only be used by MSP tooling.

To prevent future issues, there is now a check restricting the service ID length.