Fix logic for when to show add access token button
The previous logic didn't account for the allow-site-admin state so we never displayed the button in that case.
This PR fixes that.

I want to reconsider these three options in the future and see if we couldn't replace it with RBAC instead, but for now this seems like the most straightforward way to improvement.