svelte: Add more repo navigation entries and make navigation somewhat response (#59158)
This PR adds more entries to the repo nav bar, all leading to the current web app. I basically copied the same list of links we are showing on the repo page in the current app plus some additional entries (Commits and Contributors; I assume those have been removed because the repo page shows/showed a panel for them? idk)

The navbar felt really crowded, so I put the additional items into a menu. This menu will also hold the navbar items if there is not enough space. We need to make a decision about which items we want to show.

I know the "selected entry" style in the menu is not ideal. While I think in general Svelte's way of handling style is better (it prevents issues with style ordering, which we have right now), it also makes it more difficult to build somewhat styleable component libraries. I'll have to spend more time investigating this.

The includes some changes to the global navbar to make it work better on narrower screens (I originally wanted to use the same overflow approach, but the global navbar is setup differently, so I couldn't apply it as is).