search jobs: handle unknown users (#59201)
This fixes a bug where the search jobs page wouldn't render if a user
was missing. The reason is that we returned error if a user was missing,
instead of nil, so we rendered that error instead.

Test plan:
I ran a search job locally, deleted the user afterwards and verified
that the page renders as exepected.

(cherry picked from commit bdc45fa09f2e25a047675ba79b2444f1e38aa336)