grpc: forked retry package: force streaming retries to fail if we have already recieved a message on the stream (#59145)
When retrying a client stream, we must ensure that we haven't received any data from the server yet before retrying. Otherwise, we can't know if the client has already consumed part of the stream. Blindly retrying the stream could produce duplicate messages or inconsistent messages. The only safe generic behavior that we can implement is to only retry if an error occurs _before_ the server successfully sends the first message. After that, any encounters that we see on the stream will be directly returned to the caller - no retries will occur. Only the caller knows the retry semantics that it wants.


This matches the built-in grpc retry behavior (that we can't use, see https://github.com/sourcegraph/sourcegraph/issues/51060) as documented on https://learn.microsoft.com/en-us/aspnet/core/grpc/retries?view=aspnetcore-8.0#when-retries-are-valid:

> Streaming calls
> 
> Streaming calls can be used with gRPC retries, but there are important considerations when they are used together:
> 
> Server streaming, bidirectional streaming: **Streaming RPCs that return multiple messages from the server won't retry after the first message has been received. Apps must add additional logic to manually re-establish server and bidirectional streaming calls.**


As a side note: The upstream library had this behavior back in 2021 (and the discussion is a bit baffling to me): https://github.com/grpc-ecosystem/go-grpc-middleware/pull/313

## Test plan

This PR adds two additonal tests to the test suite that ensure that:

1. The library is capable of retrying the RPC if we haven't received the first message in the stream yet
2. The library will **not automatically retry** if the first message from the server has already been recieved