grpc: repo-updater: add support for automatic retries for all methods (all are idempotent) (#59130)
This PR adds support for automatic retries in the repo-updater grpc client.

---

I have gone through the repo-updater protobuf file and marked all the methods I thought were idempotent (we can't inspect this using the go protobuf packages, but I thought this was nice for documentation).

I then wrapped the basic repo-updater grpc client with an "automaticRetryClient" that uses the default retry policy that was defined in https://github.com/sourcegraph/sourcegraph/pull/59095. See that PR for more details.

CI