Searcher: fix issues with negated patterns
This refactors the searcher logic to handle negated patterns within `matcher`,
instead of the main file processing loop. This will be really helpful when
adding support for 'and' and 'or' matchers to searcher, since now negation is
handled at the level of a query atom.

The refactor uncovered two bugs in the negation logic, which are now fixed:
* When the pattern was negated, in some cases we ignored file include/ exclude patterns
* Hybrid search was totally ignoring pattern negation