msp: base startup latency alert on probe, add docstring and labels (#59314)
This change makes the startup latency alert based on `3 retries x startup probe latency`, such that the alert fires if a container fails to start up based on the startup probe. This accounts for the default Cloud Run startup probe, which is 240 seconds.

In particular, this is currently firing for the `cloud-ops` service, which is expected to take along time to start up. It uses no startup probe so it should get the 240s threshold.

---------

Co-authored-by: Dax McDonald <dax@sourcegraph.com>