search: disable structural search by default (#57584)
This disables structural search by default. As a consequence the toggle for structural search `[ ]` will be hidden in the UI and API calls with a structural search query will fail.

Users can re-enable structural search by setting `experimentalFeatures.structuralSearch: "enabled"` in site configuration.

Note: The feature is not deprecated. We are merely switching the default from "enabled" to "disabled".

Test plan:
updated unit tests: a couple of tests expected structural search to be on by default. I updated the tests to enable structural search explicitly.