Chore: clean up `RepositoryResolver` (#59402)
For reasons I don't fully understand even though it's probably my doing, RepositoryResolver depends on the search result type RepoMatch. There was some really convoluted logic about how only the name and id fields of RepoMatch were set, which is particularly confusing because a RepoMatch isn't even required to construct the RepositoryResolver. The only information being provided by RepoMatch was rev, which it really shouldn't be since a repository domain object should not even have a rev (that should be relegated to the GitCommitResolver or GitRefResolver, which contain a RepositoryResolver).

Anyways, this removes the dependency on the search result type.