search: json encode repo metadata for csv export (#59334)
We received the feedback that it is not always obvious that metadata is line-separated in the CSV export of search results. Depending on the tool used to display the CSV, only the first metadata value might be visible. Additionally the current format can be hard to parse. 

Here we add a new column "Repository metadata JSON" to the CSV export of search results, which encodes metadata into a single line JSON. We deprecate "Repository metadata" but keep it around for backward compatibility for short period of time. 

## Test plan
- updated unit test
- manual testing: 
  - validated the JSON encoding with a python script that reads the CSV and parses the JSON field.