db: Cast some slices of api.RepoID and api.RepoName to primitive slices
When using primitive slices, the pq implementation is more optimized, and we also get better handling in tracing for the maximum length.

As mentioned on Slack here: https://sourcegraph.slack.com/archives/C02UC4WUX1Q/p1704928801529719?thread_ts=1704913928.968849&cid=C02UC4WUX1Q, this is not an ultimate solution. It fixes it for some of our queries, but won't be a sustainable long-term solution.

Alternative ideas:
- Remove these two type aliases
- Fork the pq.Array implementation and add native support for those two

## Test plan

CI should find issues with the DB serialization.