search jobs: expose fromMatch
This is a pure refactor.

I move `fromMatch` and all related code from `frontend/internal` to
`/internal` because I want to share the type converters of the Stream
API with Search Jobs. The goal is that Search Jobs and the Stream API
offer the same API with respect to matches.

Test plan:
CI