Symbol search: clean up job logic (#59500)
The search planning code uses the same `TextPatternInfo` struct for unindexed
search, structural search, and symbol search. However, most of the parameters
don't make sense for symbol search. This PR pulls out a dedicated
`SymbolSearchRequest` struct for symbol search to use instead.

This uncovered an error with symbol search where negation was allowed, but gave
the wrong results. With this change, we just return an error.