msp/monitoring: add external uptime check and alert, rework health probes configuration (#59461)
The new configuration is mostly based on Cody Gateway - if a service has an external domain, we create an uptime check and alert on failures.

The uptime check uses MSP standards, which depends on whether or not service health probes are configured. Since we use this in several places now, I've also reworked the health probes configuration to make it easier to reason with:

1. `healthProbes` now configures all healthchecks. `startupProbe` and `livenessProbe` has been removed
2. `disabled` is now `healthzProbes` - this configures if MSP healthchecks should be used, instead of default `/` ones.
3. By default, if no config is provided, MSP healthchecks are not used
4. If config is provided, MSP healthchecks must be explicitly disabled

Closes https://github.com/sourcegraph/managed-services/issues/350

This is required for our upcoming vendor evaluations as well.

This PR also includes a variety of internal improvements to alert policies.