Search: add language to the `FileMatch` type (#59616)
The language of a file cannot reliably be determined with just the file name. We have numerous examples of this failing and giving surprising results.

We can provide much more reliable language detection if we have the file contents available to provide more information during language detection. However, fetching the contents of files to detect the language can be expensive, especially for operations like search where there may be many results we would need to fetch contents for.

Instead, this adds the language as a field on the FileMatch result. We already get this information from Zoekt and symbols, and it will be straightforward to add language detection in searcher, where the file contents are already available when constructing matches.