Search: revert automatic retries (#59603)
This reverts #59133, which added automatic retries for the gRPC client that
Zoekt uses.

As noted in #59133 (comment), this is a change in behavior from before. We
suspect that when Zoekt is overloaded, it can appear as "unavailable", which
causes us to automatically retry searches. This in turn puts more load on
Zoekt, keeping it in an overloaded state. In general, Zoekt has special
handling for when replicas are unavailable, and we don't want to use this
general default.