Search: always show keyword search toggle (#59610)
Currently when the old "search input" is enabled, the "Precise (NEW) 💖" option
doesn't show up in the toggle. Now we make sure to pass the feature flag value
to the old components, so the option correctly shows up.

The behavior was extra confusing because we still used the new search mode, but
didn't show that it was activated!