search jobs: switch from CSV to JSON (#59619)
Relates to #59352 

We switch the format of the results download from CSV to line-separated JSON. 
Each line corresponds to a JSON object containing chunk matches. 
The JSON object has the same format as the matches served by the Stream API.

This is a **breaking change** motivated by customer feedback. 
Search Jobs is still released as EAP so a breaking change is acceptable. 

Pros:
- richer information (matches, content, positions)
- supports all result types
- same format as Stream API

Cons:
- Requires more storage
- Not as easy to parse by a human as a CSV

## Test plan
- updated and new units tests