Searcher: natively support AND/ OR patterns (#59571)
This is the final PR to handle AND/ OR queries directly in searcher, instead of
running each clause separately and combining the results in frontend. Major
changes:
1. Update `job.go` to pass entire query to searcher, instead of creating flat
jobs
2. Update `TextPatternInfo` and `protocol.PatternInfo` to hold the whole query
instead of the string pattern directly. This was tricky, since these types are
also used for structural search, which does not support AND/ OR directly.
3. Within searcher, convert this query to a `matchTree`