feat: sanitize generated HTML
Common bluemonday.Policy defined in a shared package,
which presents a convenient API for working with the 3 main functions.
Key policy rules tested in policy_test.go to prevent regressions.

Switched from .Sanitize to .SanitizeReader in markdown.go:
bluemonday always uses the latter under the hood,
so we just cut to the chase and omit double conversion.

#57797