gomod: update zoekt
https://github.com/sourcegraph/zoekt/compare/7487a0d53131...b82a981f8240

There are quite a few changes, but most are internal e2e test changes.
Here are the commits which change behaviour. Only the scoring change
will have a minor effect on users in rare cases:

- 7e44ea78f6 all: adjust field order for match structs
- 7dcc7979f6 api: implement succinct output for SearchOptions.String
- 1adab43b4c score: remove ZOEKT_NOVELTY_DISABLE
- 18b6999367 api: JSON omitempty on optional fields in FileMatch
- b82a981f82 score: do scoring on candidateMatches

And below are the e2e test changes:

- 155050eb98 zoekt-archive-index: split out ranking tests and archive indexing
- 344e33ae3c e2e: introduce optional shard cache
- 6463096a03 e2e: recall and visual indication of ranks
- b35d8a202d e2e: add ranking tests for atom boosting
- 4be2929fd0 e2e: add a ranking test case for "zoekt searcher"

Test Plan: tested in zoekt repo and this CI