search: remove check for structural search from backend (#59696)
search: remove check for structural search

This removes the check for structural search from the backend.  This
means, even if the `[ ]`-toggle is hidden in the UI, our backend will
still support structural search.

The intend is to give customers a few releases to either explicitly
enable structural search or to migrate their links, code insights and
scripts to regexp.

Test plan:
I tested manually that structural search queries work if the pattern
type is specified explicitly in the search bar.