Perf: cache whether subrepo perms are enabled for a repo (#59713)
Currently, in the SubRepoPermsJob, we filter each search event in the stream serially. This is problematic because these checks require a database call, which means we're making a serial database call for each unique repo in our search results. We pay this cost whether or not the repo is a perforce repo because we're making the DB call specifically to check whether the repo is a perforce repo. This is currently adding ~2ms per unique repo in a set of search results when subrepo permissions are enabled (such as on S2).

We had various layers of caching of this information that I consolidated into a cache on the checker itself so the caller doesn't need to worry about cache logic and so the cache can be shared between dependents of a client. Whether a repo is a perforce repo is likely to be quite stable, so I set a default cache TTL of 1 hour. This can realistically probably be higher, but I'd rather err on the side of caution since this is a security feature. The cache itself is very dense since it uses roaring bitmaps. I am not concerned about memory usage given that we already have a much larger cache for the actual rules.

Since the cache lives on the client, it persists for the lifetime of the client. In most cases, this is the lifetime of the process since we use the global DefaultSubRepoPermsClient in most cases.