gitserver: Simplify client
This PR consolidates and simplifies the client code a lot, in anticipation of moving more of the logic to the server-side.
This allows us to have a better overview of what is actually used, and keeps the client interface as simple as possible.

Most notably, the batchLog endpoint is no longer used.
It was used for a single use-case and added quite a lot of complexity. My expectation is that with the move to server-side commit logging and parsing, we will be able to make up for the batch log performance improvement.

## Test plan

CI still passes, including integration tests.