Search contexts: paginate repo names (#58685)
A customer reported that search contexts with over a thousand repos were failing when trying to create them.

The root cause is that the webapp was fetching repositories to validate them, but was not using pagination, and instead tried to get them all in one go. The backend limits the number of results in a page to 1000, even if the client requests more. Now, the web app pages through results.

(cherry picked from commit 439728733eb1fb080164dafeb35dc6ba9edc5948)