Search: Add initial keyword search toggle (#59779)
This change adds a simple toggle to enable/ disable keyword search, and removes
the prototype toggle called "Precise (NEW)". It also removes the smart search
toggle, since smart search doesn't pair well with the new keyword search
behavior. When the feature is disabled, we revert to the old UI and defaults
(smart search is still present, and we default to the 'standard' patterntype).
This helps mitigate risk for the initial release, since we can always disable
the change completely.

The toggle behaves like other toggles in the nav bar: the setting sticks unless
the user opens a new tab or refreshes the page. Then it reverts back,
defaulting to "keyword search".